BaseTenAppKit is now 64-bit, too, and create-universal-binaries.sh works again
git-svn-id: http://basetenframework.org/svn/trunk@1330 fe1aa653-7022-0410-af67-bafd91c7e665

+++ BaseTenAppKit/Sources/BXAuthenticationPanel.m
@@ -55,7 +55,7 @@ static const CGFloat kSizeDiff = 25.0;
 									  backing: NSBackingStoreBuffered defer: YES] autorelease];
 }
 
-- (id) initWithContentRect: (NSRect) contentRect styleMask: (unsigned int) styleMask
+- (id) initWithContentRect: (NSRect) contentRect styleMask: (NSUInteger) styleMask
                    backing: (NSBackingStoreType) bufferingType defer: (BOOL) deferCreation
 {
     if ((self = [super initWithContentRect: contentRect styleMask: styleMask
